Various juicy goodness: new status indicators, camera-aligned target reticles, some cleanup of RenderUtil and commonly used render states. The status indicators don't yet display damage or rotate to display the useful information toward the user regardless of camera location.

+++ src/java/com/threerings/bang/client/BangMetrics.java
@@ -12,5 +12,5 @@ public class BangMetrics
     public static final int SQUARE = 40;
 
     /** The dimensions of a single board tile. */
-    public static final int TILE_SIZE = 10;
+    public static final float TILE_SIZE = 10;
 }
